[PM-11689] Refactoring implementation to better unify logic
